+++ multiplexer/modeling/proposal_generator/spn.py
@@ -523,21 +523,21 @@ class SEGPostProcessor(torch.nn.Module):
         only for helping principle understand.
         """
         mask = np.zeros_like(bitmap, dtype=np.uint8)
-        cv2.fillPoly(mask, box.reshape(1, 4, 2).astype(np.int32), 1)
+        cv2.fillPoly(mask, box.reshape(1, 4, 2).astype(int), 1)
         return cv2.mean(bitmap, mask)[0]
 
     def box_score_fast(self, bitmap, _box):
         h, w = bitmap.shape[:2]
         box = _box.copy()
-        xmin = np.clip(np.floor(box[:, 0].min()).astype(np.int), 0, w - 1)
-        xmax = np.clip(np.ceil(box[:, 0].max()).astype(np.int), 0, w - 1)
-        ymin = np.clip(np.floor(box[:, 1].min()).astype(np.int), 0, h - 1)
-        ymax = np.clip(np.ceil(box[:, 1].max()).astype(np.int), 0, h - 1)
+        xmin = np.clip(np.floor(box[:, 0].min()).astype(int), 0, w - 1)
+        xmax = np.clip(np.ceil(box[:, 0].max()).astype(int), 0, w - 1)
+        ymin = np.clip(np.floor(box[:, 1].min()).astype(int), 0, h - 1)
+        ymax = np.clip(np.ceil(box[:, 1].max()).astype(int), 0, h - 1)
 
         mask = np.zeros((ymax - ymin + 1, xmax - xmin + 1), dtype=np.uint8)
         box[:, 0] = box[:, 0] - xmin
         box[:, 1] = box[:, 1] - ymin
-        cv2.fillPoly(mask, box.reshape(1, 4, 2).astype(np.int32), 1)
+        cv2.fillPoly(mask, box.reshape(1, 4, 2).astype(int), 1)
         return cv2.mean(bitmap[ymin : ymax + 1, xmin : xmax + 1], mask)[0]
